Add settings.js file for settings
This is part of the master plan (mu-hahahahaha) to make it easy to
configure parts of the Brigade UI via the chart. It moves settings into
a designated file in its own directory. Inside of the chart, we will be
able to write a secret or configmap to store these settings, and then
parameterize them in values.yaml.