fix(*): fix several minor things for docker builds
There were a few minor issues that were preventing a Docker build from
working with the latest codebase. Mostly, they were related to the
differences between macOS and Linux.

I added back the `yarn docker-build` target that is documented in the
installation instructions, but which seems to have gotten lost on this
branch.